requirements.txt
yfinance
pandas
pyspark==3.4.4
delta-spark==2.4.0
setuptools
streamlit
plotly 